modify: CommentSection -> 공통 컴포넌트로 빼서 common/comment 경로에 CommentSection, CommentItem, CommentMenu로 분리 후 '더보기'버튼 팝업 기능 구현
